requirements.txt
discord.py==1.4.1
more-itertools==8.4.0
aioitertools==0.7.0